Issue #204 Move taxonomy module so that is can be easily accesses by both human_services and newcomers' guide modules
