+++ human_services/services_at_location/tests/test_api.py
@@ -4,8 +4,8 @@ from human_services.locations.tests.helpers import LocationBuilder
 from human_services.services_at_location.tests.helpers import ServiceAtLocationBuilder
 from human_services.organizations.tests.helpers import OrganizationBuilder
 from human_services.services.tests.helpers import ServiceBuilder
-from human_services.taxonomies.tests.helpers import TaxonomyTermBuilder
 from human_services.locations.models import ServiceAtLocation
+from taxonomies.tests.helpers import TaxonomyTermBuilder
 
 
 class ServicesAtLocationApiTests(rest_test.APITestCase):
