+++ taxonomies/apps.py
@@ -3,5 +3,5 @@ from django.utils.translation import ugettext_lazy as _
 
 
 class TaxonomiesConfig(AppConfig):
-    name = 'human_services.taxonomies'
+    name = 'taxonomies'
     verbose_name = _("Taxonomies")
